requirements.txt
azure-ai-textanalytics==5.1.0
azure-common==1.1.27
azure-core==1.21.1
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
isodate==0.6.1
msrest==0.6.21
numpy==1.21.5
oauthlib==3.1.1
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
